requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.13
whitenoise==6.2.0